refactor: move code and tests for verification into dedicated files
